# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Camera {
  id: string;
  name: string;
  /** @nullable */
  rtsp_url?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  enabled?: boolean;
  /** @nullable */
  resolution?: string | null;
  /** @nullable */
  fps?: number | null;
  /** @nullable */
  codec?: string | null;
}

export interface CameraInput {
  id: string;
  name: string;
  rtsp_url?: string;
  location?: string;
  lat?: number;
  lng?: number;
  enabled?: boolean;
  resolution?: string;
  fps?: number;
  codec?: string;
}

export interface CameraListResponse {
  success: boolean;
  count: number;
  cameras: Camera[];
}

export interface CameraScanResult {
  success: boolean;
  /** @nullable */
  rtsp_url?: string | null;
}

export interface SimpleResponse {
  success: boolean;
  message?: string;
  error?: string;
}

export interface DetectionAttribute {
  attribute_type?: string;
  attribute_value?: string;
  confidence?: number;
}

export interface Incident {
  id: number;
  event_id: number;
  object_type: string;
  confidence: number;
  status: string;
  timestamp: string;
  /** @nullable */
  camera_id?: string | null;
  /** @nullable */
  severity?: string | null;
  /** @nullable */
  source_description?: string | null;
  attributes?: DetectionAttribute[];
}

export interface IncidentListResponse {
  success: boolean;
  count: number;
  results?: Incident[];
  incidents?: Incident[];
  statuses?: string[];
}

export type IncidentStatusInputStatus = typeof IncidentStatusInputStatus[keyof typeof IncidentStatusInputStatus];


export const IncidentStatusInputStatus = {
  New: 'New',
  Acknowledged: 'Acknowledged',
  In_Progress: 'In Progress',
  Resolved: 'Resolved',
  False_Alarm: 'False Alarm',
} as const;

export interface IncidentStatusInput {
  status: IncidentStatusInputStatus;
}

export interface IncidentStatusResult {
  success: boolean;
  event_id: number;
  status: string;
}

export type PaypalOrderInputContacts = { [key: string]: unknown };

export interface PaypalOrderInput {
  planId: string;
  planName?: string;
  amount?: string;
  currency?: string;
  district?: string;
  contacts?: PaypalOrderInputContacts;
}

export type PaypalOrderResultOrder = { [key: string]: unknown };

export interface PaypalOrderResult {
  success: boolean;
  id: string;
  status: string;
  order?: PaypalOrderResultOrder;
}

export interface PaypalCaptureInput {
  planId?: string;
  orderId?: string;
}

export type PaypalCaptureResultCapture = { [key: string]: unknown };

export interface PaypalCaptureResult {
  success: boolean;
  status: string;
  capture?: PaypalCaptureResultCapture;
}

export type DeleteCameraParams = {
id: string;
};

export type SearchAttributesParams = {
object_type?: string;
color?: string;
equipment?: string;
camera_id?: string;
min_confidence?: string;
start_date?: string;
end_date?: string;
};