# lib/api-zod/src/generated/types/cameraInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CameraInput {
  id: string;
  name: string;
  rtsp_url?: string;
  location?: string;
  lat?: number;
  lng?: number;
  enabled?: boolean;
  resolution?: string;
  fps?: number;
  codec?: string;
}