# lib/api-zod/src/generated/types/cameraListResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Camera } from './camera';

export interface CameraListResponse {
  success: boolean;
  count: number;
  cameras: Camera[];
}