# lib/api-zod/src/generated/types/incidentListResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Incident } from './incident';

export interface IncidentListResponse {
  success: boolean;
  count: number;
  results?: Incident[];
  incidents?: Incident[];
  statuses?: string[];
}