# lib/api-zod/src/generated/types/incidentStatusInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { IncidentStatusInputStatus } from './incidentStatusInputStatus';

export interface IncidentStatusInput {
  status: IncidentStatusInputStatus;
}