# lib/api-zod/src/generated/types/incidentStatusResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface IncidentStatusResult {
  success: boolean;
  event_id: number;
  status: string;
}