# lib/api-zod/src/generated/types/paypalCaptureResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaypalCaptureResultCapture } from './paypalCaptureResultCapture';

export interface PaypalCaptureResult {
  success: boolean;
  status: string;
  capture?: PaypalCaptureResultCapture;
}