# lib/api-zod/src/generated/types/paypalOrderInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaypalOrderInputContacts } from './paypalOrderInputContacts';

export interface PaypalOrderInput {
  planId: string;
  planName?: string;
  amount?: string;
  currency?: string;
  district?: string;
  contacts?: PaypalOrderInputContacts;
}