# lib/api-zod/src/generated/types/paypalOrderResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaypalOrderResultOrder } from './paypalOrderResultOrder';

export interface PaypalOrderResult {
  success: boolean;
  id: string;
  status: string;
  order?: PaypalOrderResultOrder;
}