# lib/api-zod/src/generated/types/simpleResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SimpleResponse {
  success: boolean;
  message?: string;
  error?: string;
}